IC/Relay/requirements.txt
aiofiles==24.1.0
asyncio==3.4.3
paho-mqtt==2.1.0
pycryptodome==3.20.0
websockets==13.0.1
pandas==2.2.3
bluepy==1.3.0
crc==7.0.0